Encapsulation of power-up behaviour in Player class
Methods and properties related to power-ups in Player have been moved to the new script PlayerPowerupManager.cs
