+++ Assets/Resources/Scripts/Player/PlayerPowerupManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81addf6a93e7a49eaa0d3b9a22dc03c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
